Move ExplorerWeb-only factory helpers to ExplorerWeb.Factory
