+++ apps/explorer/test/support/factory.ex
@@ -1,8 +1,6 @@
 defmodule Explorer.Factory do
   use ExMachina.Ecto, repo: Explorer.Repo
 
-  import Ecto.Query
-
   alias Explorer.Chain.{Address, Block, Hash, InternalTransaction, Log, Receipt, Transaction}
   alias Explorer.Repo
 
@@ -128,32 +126,6 @@ defmodule Explorer.Factory do
     insert(:receipt, transaction_hash: hash, transaction_index: index)
 
     Repo.preload(block_transaction, [:block, :receipt])
-  end
-
-  def with_block(%Transaction{index: nil} = transaction, %Block{hash: block_hash}) do
-    next_transaction_index = block_hash_to_next_transaction_index(block_hash)
-
-    transaction
-    |> Transaction.changeset(%{block_hash: block_hash, index: next_transaction_index})
-    |> Repo.update!()
-    |> Repo.preload(:block)
-  end
-
-  def with_receipt(%Transaction{hash: hash, index: index} = transaction) do
-    insert(:receipt, transaction_hash: hash, transaction_index: index)
-
-    Repo.preload(transaction, :receipt)
-  end
-
-  defp block_hash_to_next_transaction_index(block_hash) do
-    query =
-      from(
-        transaction in Transaction,
-        select: transaction.index,
-        where: transaction.block_hash == ^block_hash
-      )
-
-    Repo.one!(query) + 1
   end
 
   defp integer_to_hexadecimal(integer) when is_integer(integer) do
